requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
python-dotenv==1.0.1
Werkzeug==3.0.1
SQLAlchemy>=2.0.16
click>=8.1.3
itsdangerous>=2.1.2
Jinja2>=3.0.1
MarkupSafe>=2.0.1
email-validator==1.1.3 